msstor_cache: added ms speed up support

+++ Satelite/clock.c
@@ -55,7 +55,7 @@ int bus2no(int cpu)
 void change_clock(int dir, int flag)
 {
 	int sel;
-	int *cpu[2];
+	u8 *cpu[2];
 
 	if(flag) {
 		cpu[0]=&(cnf.umdisocpuspeed);
@@ -110,7 +110,7 @@ void change_region(int dir, int max)
 void change_plugins(int dir, int flag)
 {
 	int sel;
-	int *plugins;
+	u8 *plugins;
 
 	if(flag == 0) {
 		plugins=&(cnf.plugvsh);
